# packages/server/src/api/public/docs/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "appName API",
		"description": "Official backend surface for the appName media experience. Routes are grouped under `/v1/api` for client-facing functions and `/v1/admin` for back-office operations.",
		"version": "1.0.0",
		"contact": {
			"name": "App Support",
			"url": "https://appName.io",
			"email": "support@appName.io"
		}
	},
	"servers": [
		{
			"url": "https://api.appName.io/v1/api",
			"description": "Production"
		},
		{
			"url": "http://localhost:3000/v1/api",
			"description": "Local development"
		}
	],
...
	"paths": {
...
		"/v1/api/subscription/paypal/create/{planId}": {
			"post": {
				"tags": [
					"Subscription"
				],
				"summary": "Create PayPal subscription",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "planId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"redirectURI",
									"cancelURI"
								],
								"properties": {
									"redirectURI": {
										"type": "string",
										"format": "uri"
									},
									"cancelURI": {
										"type": "string",
										"format": "uri"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Subscription created or approval URL returned",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/HttpSuccess"
										},
										{
											"type": "object",
											"properties": {
												"data": {
													"type": "object",
													"properties": {
														"url": {
															"type": "string",
															"format": "uri",
															"description": "PayPal approval URL or frontend redirect URL"
														}
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request or subscription creation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpError"
								}
							}
						}
					},
					"404": {
						"description": "User or plan not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpError"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/api/subscription/paypal/one-time/create/{planId}": {
			"post": {
				"tags": [
					"Subscription"
				],
				"summary": "Create a one-time PayPal payment order",
				"description": "Creates a PayPal Order for a one-time payment using the Orders v2 API. The user pays once and gets 30 days of access — no recurring billing. A local Subscription record (CREATED status) is persisted before returning the approval link. After the user approves on PayPal, the existing capture-subscription endpoint (`GET /subscription/capture`) will capture and activate the subscription via revalidation.",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "planId",
						"in": "path",
						"required": true,
						"description": "Public ID of the plan to subscribe to",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"redirectURI",
									"cancelURI"
								],
								"properties": {
									"redirectURI": {
										"type": "string",
										"format": "uri",
										"description": "URL to redirect the user to after payment approval"
									},
									"cancelURI": {
										"type": "string",
										"format": "uri",
										"description": "URL to redirect the user to if they cancel the payment"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "PayPal approval URL returned",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/HttpSuccess"
										},
										{
											"type": "object",
											"properties": {
												"data": {
													"type": "object",
													"properties": {
														"url": {
															"type": "string",
															"format": "uri",
															"description": "PayPal approval URL — redirect the user here to complete payment"
														}
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request, subscription creation failed, or user already subscribed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpError"
								}
							}
						}
					},
					"404": {
						"description": "User or plan not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
...
		"schemas": {
			"HttpSuccess": {
				"type": "object",
				"required": [
					"success",
					"message"
				],
				"properties": {
					"success": {
						"type": "boolean",
						"example": true
					},
					"message": {
						"type": "string"
					},
					"data": {
						"description": "Response data (optional)"
					}
				}
			},
			"HttpError": {
				"type": "object",
				"required": [
					"code",
					"message"
				],
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"description": "Additional error details (optional)"
					}
				}
			},
...
		}
	}
}